.windsurf/rules/requirements.txt
# Core Dependencies
pyyaml>=6.0.1
psutil>=5.9.5

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1

# Testing Dependencies
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
